[statsdreceiver] support simple tags
dogstatsd supports two types of tags on metrics: simple and
dimensional tags[^1]. the former is just a key, the latter is
a key and a value.

with the assumption that many users of the statsdreceiver are
enabling ingest of dogstatsd metrics, this makes the statsd
parser more optimistic, so it can handle tags w/o a value.

when this happens, we set an attribute that has a zero value.
so far as i know, this is allowed in the opentelemetry spec.
the decision of how to handle attributes w/ zero values is
best left to configuration w/in the pipeline itself, as different
users may have different opinions or approaches that work best
with their systems.

[^1]: https://www.datadoghq.com/blog/the-power-of-tagged-metrics/#whats-a-metric-tag