[statsdreceiver] add datadog_summary observer
this adds a new observer that produces a "summary" of histograms
similar to how datadog does it. this is not a distinct metric
type in datadog, but instead is a convention where a sum and multiple
gauges are emitted for the incoming histogram. the sum tracks the
total number of samples and the gauges track mean, median, sum, and
p95 aggregations over the sample set.

the primary use-case for this observer is admittedly niche; it is
designed for cases where you are substituting the OpenTelemetry
Collector in place of the Datadog Agent for DogStatsD ingest, but
don't want to break existing dependencies on existing metrics that
aggregate over histograms.