add explicit histogram observer
not all backends currently support storing exponentialhistograms, as they
are pretty new to otel. this adds a new histogram observer that aggregates
to an otel Histogram with explicit boundaries equal to the boundaries that
would be selected by the exponential histogram algorithm, assuming scale=0
and an assumed bounds of 2^-15...2^15. in the case of a latency distribution
this can handle values from 0.03ms to ~9 hours.