make distribution and histogram distinct types
there is a subtle difference between the Histogram and Distribution
type; the Distribution is specific to DogStatsD and performs global
aggregation by taking in raw values. the Histogram performs local
(node) aggregation and outputs preaggregated avg/count/max/median
and optionally percentile aggregations.

this updates the receiver to treat these as explicit types, which must
be configured separately. prior, there was behavior where the same
config would be applied to each statsd type even if the end-user desired
them to differ.